+++ server/routes.js
@@ -1,0 +1,12 @@
+
+module.exports = function(server){
+
+  server.route({
+    method: 'GET',
+    path: '/{name}',
+    handler: (request, reply) => {
+      reply(`hello ${request.params.name}`)
+    }
+  });
+  
+};
